% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{avgRandomNetMet}
\alias{avgRandomNetMet}
\title{Average Metrics from Randomly Generated Networks}
\usage{
avgRandomNetMet(n, p)
}
\arguments{
\item{n}{Number of nodes}

\item{p}{Probability of link formation for any pair of nodes}
}
\value{
A column of mean metrics for 40 randomly generated networks
}
\description{
Executing this function will randomly generate 40 networks, calculate
metrics, store the metrics as columns in a temporary matrix, calculate
the means of each row of the matrix, and return the mean metrics.
}
\examples{
ravgmet <- avgRandomNetMet(40,0.32)
}
