% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{avgSynthNetMet}
\alias{avgSynthNetMet}
\title{Mean Metrics of Synthesized Social Networks from a matrix of personalities}
\usage{
avgSynthNetMet(team, ctbl, netdensity)
}
\arguments{
\item{team}{A matrix that identifies a set of personality types}

\item{ct}{A matrix that is a personality compatibility look up table}

\item{d}{A desired density for the synthesized social networks}
}
\value{
A column of mean metrics for 40 synthsized social networks
}
\description{
Given an assignment of personalities and a compatibility table,
this function will synthesize 40 social networks and calculate the
mean metrics. This function receives a single column matrix of
personalities types and a personality compatibility look up table.
}
\examples{
group <- replicate(10,makeMBTI())
cpat <- read.csv("CustomCompatibilityTable.csv")
savgmet <- avgSynthNetMet(group,cpat,0.32)
}
