% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\docType{data}
\name{cachet}
\alias{cachet}
\title{Establish an Environment}
\format{An object of class \code{environment} of length 3.}
\usage{
cachet
}
\description{
Distinguishing mark or seal is one of the definitions of the wor cachet.
As an environment within this package, cachet will provide a place to
assign global variables within this package's functions.
}
\examples{
assign("evaluations",lookups, envir=cachet) # Notice the quotes.
assign("rngSeed",.Random.seed,envir=cachet) # Store current random seed
}
\keyword{datasets}
