% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{calculateMetrics}
\alias{calculateMetrics}
\title{Calculate Metrics}
\usage{
calculateMetrics(socialnetwork)
}
\arguments{
\item{socialnetwork}{An igraph object}
}
\value{
A column of metrics
}
\description{
Dozens of metrics are available in the igraph package. This function
accepts an igraph object and returns a column of metrics for use by
the comparison table generator function in this package.
}
\examples{
sn <- realnetwork("sociomatrix.txt")   # Read sociomatrix to make a network.
sn.metrics <- calculateMetrics(sn)     # Calculate a set of metrics.
}
