% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{compileReport}
\alias{compileReport}
\title{Compile a Report}
\usage{
compileReport(rwmet, mmet1, mmet2)
}
\arguments{
\item{rwmet}{Real world network metrics}

\item{mmet1}{Mean metrics of a population of randomly generated networks.}

\item{mmet2}{Mean metrics for compatibility synthesized network population.}
}
\value{
A report suitable for the report table generator.
}
\description{
After calculating the metrics of a real world network, means of metrics
for a population of randomly generated networks, and means of metrics for
a population of personality-based synthesized networks, call this function
to compile the results into a report. The data returned by this function
can be used by the report table generator in this package.
}
\examples{
sn <- realnetwork("sociomatrix.txt")  # Metrics from a real world network
ravgmet <- avgRandomNetMet(40,0.32)   # Mean metrics from random net pop.
savgmet <- socialnetSynthesizer(40,0.32) # Mean metrics
rpt <- compileReport(sn,ravgmet,savgmet)
}
