% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{compileSearchComparisonReport}
\alias{compileSearchComparisonReport}
\title{Compile Algorithm Comparison Report}
\usage{
compileSearchComparisonReport(fnMC, fnGA, reportTitle)
}
\arguments{
\item{fnMC}{A file name of a Monte Carlo search results report}

\item{fnGA}{A file name of a Genetic Algorithm search results report}
}
\value{
HTML code that can be written to a file
}
\description{
After generating reports that compare the search algorithms with the
randomly generated networks, this function compiles a data.frame that
compares the metrics resulting from personality assignments found by
the two search algorithms.
}
\examples{
fnMC <- "../reports/mcRawReport_RobinsBank.txt"
fnGA <- "../reports/gaWeaselRawReport_RobinsBank.txt"
compileSearchComparisonReport(fnMC,fnGA)
}
