% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snMetricsEvolver.r
\name{gaMetricsSearch}
\alias{gaMetricsSearch}
\title{Genetic Algorithm Search for Optimum Network Metrics}
\usage{
gaMetricsSearch(iSampleSize, iPopSize, fnMat, fnMet, ct, fnStat, fnOpt, fnR,
  fnS, tLimit)
}
\arguments{
\item{ct}{A personality compatibility look up table}

\item{fnStat}{A file name for a status report from this search function}

\item{fnOpt}{A text file with a matrix with the optimum set of personalities}

\item{rwmat}{A real world sociomatrix}
}
\value{
This function does not return anything
}
\description{
Using a real world sociomatrix for comparison, this evolutionary 
algorithm produces an intial population of sociomatrices and compares
the metrics with the metrics of network generated from the real world
sociomatrix. A fitness function compares the difference between the
metrics and returns a score.
}
\examples{
cpat <- as.matrix(read.csv("CompatibilityTable.csv")
rbmet <- calculateMetrics(rb)      
rgmet <- avgRandomNetMet(rbmet[1],rbmet[4])
gaSearch(1000,"Robins_Bank.txt",cpat,"gaStat.csv","GA_OptSet.csv")
}
