% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{generateComparisonTable}
\alias{generateComparisonTable}
\title{Generate an Algorithm Comparison Table}
\usage{
generateComparisonTable(rpt, reportTitle)
}
\arguments{
\item{rpt}{A compiled report with five columns of metrics}

\item{reportTitle}{The name of the real world social network}
}
\value{
HTML code that can be written to a file
}
\description{
The report generator accepts a compiled report, which is produced by
another function in this package. This function uses the formatter and
format_table functions from the formattable package. The formatted table
returned from this function can be written to an HTML file and viewed
with a web browser.
}
\examples{
sn <- realnetwork("sociomatrix.txt")  # Metrics from a real world network
ravgmet <- avgRandomNetMet(40,0.32)   # Mean metrics from random net pop.
savgmet <- socialnetSynthesizer(40,0.32) # Mean metrics
rpt <- compileReport(sn,ravgmet,savgmet)
dfrmTbl <- generateTable(rpt,"Social Network Name")
fileConn<-file("outputfileName.html")
writeLines(dfrmTbl, fileConn)
close(fileConn)
}
