% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{makeMBTI}
\alias{makeMBTI}
\title{Make an Myers Briggs Type Indicator}
\usage{
makeMBTI()
}
\description{
This function constructs an Myers Briggs Type Indicator (MBTI) in
accordance with the following table:
Extroverts 46.3%	or Introverts 53.7%
iNtuitive  31.9%	or Sensing    68.1%
Thinking   52.9%	or Feeling    47.1%
Judging    58.1%	or Perceiving 41.9%
Hammer, Allen L., and Wayne D. Mitchell. "The distribution of MBTI types
in the US by gender and ethnic group." Journal of Psychological Type 37
(1996): 2-15.
}
