% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{makeSociomatrix}
\alias{makeSociomatrix}
\title{Make a Sociomatrix}
\usage{
makeSociomatrix(a, b, lut, goaldensity)
}
\arguments{
\item{a}{list of indices for the compatibility table}

\item{b}{list of indices for the compatibility table}

\item{lut}{a compatibility table}
}
\value{
a matrix
}
\description{
Based upon random numbers and a compatibility table,
this function makes a symmetric sociomatrix.
}
\examples{
comptable <- as.matrix(read.csv(ctableName))
team <- read.csv("personalityList.csv", header=FALSE)
indices <- match(team,colnames(comptable)
sm <- makeSociomatrix(indices,indices,comptable)
}
