% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{mcSearch}
\alias{mcSearch}
\title{Monte Carlo Search for Optimum Personality Assignment}
\usage{
mcSearch(iSampleSize, fnMat, fnMet, ct, fnStat, fnOpt, fnR, fnS, tLimit)
}
\arguments{
\item{iSampleSize}{Number of sociomatrices to synthesize for statistics}

\item{fnMat}{Input real world sociomatrix for comparison.}

\item{fnMet}{file name for the best metrics found during this search}

\item{fnStat}{A file name for a status report from this search function}

\item{fnOpt}{A text file with a matrix with the optimum set of personalities}

\item{tLimit}{A time limit in number of seconds}

\item{lut}{A personality compatibility look up table}
}
\value{
This function does not return anything
}
\description{
Using real world and mean metrics of randomly generated networks,
this function synthesizes a population of social networks, averages
the metrics, and compares the mean metrics with the metrics from
the real world network and the randomly generated network population.
Results are compiled into a report and a report generator produces
an HTML file with a color coded table and a status report of the search.
}
\examples{
cpat <- as.matrix(read.csv("CompatibilityTable.csv")
rb <- realnetwork("RobinsBank.txt")  # Generate an igraph object
rbmet <- calculateMetrics(rb)
rgmet <- avgRandomNetMet(rbmet[1],rbmet[4])
mcSearch(rbmet,rgmet,cpat,"MC_SearchStatus.csv","OptimumSet.csv")
}
