% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{meanSynthNetMet}
\alias{meanSynthNetMet}
\title{Mean Metrics of Synthesized Social Networks from a file of personalities}
\usage{
meanSynthNetMet(fn, fnct, d)
}
\arguments{
\item{fn}{File name of a file that contains an assignment of personalities}

\item{fnct}{A file that contains personality compatibility table}

\item{d}{A desired density for the synthesized social networks}
}
\value{
A column of mean metrics for 30 synthsized social networks
}
\description{
Given an assignment of personalities and a compatibility table,
this function will synthesize 30 social networks and calculate the
mean metrics. This function reads a file that contains an assignment of
personalities; so, the intent of this function is to work with
the optimized assignment.
}
\examples{
ravgmet <- meanSynthNetMet("personalities.csv","compatibilities.csv",0.32)
}
