% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{realnetwork}
\alias{realnetwork}
\title{A Real Network}
\usage{
realnetwork(fname)
}
\arguments{
\item{fname}{Name of a text file containing a sociomatrix.}
}
\value{
An iGraph object
}
\description{
A sociomatrix is a square matrix where each row and column represent
a person.  A one in a cell at the intersection of a row and column
indicates that the two people have some connection.  A zero in a cell
indicates that the two people do not have a connection. This function
reads a sociomatrix and returns a network as an igraph object.
}
\examples{
sn <- realnetwork("sociomatrix.txt")
}
