% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynthesizedNetworkComparisonUtilities.r
\name{setMetricGoals}
\alias{setMetricGoals}
\title{Set Goals for the Metrics}
\usage{
setMetricGoals(metA, metB)
}
\arguments{
\item{metA}{a set of metrics for a real world network}

\item{metB}{a set of metrics for a random or synthesized network}
}
\description{
setMetricGoals receives two sets of metrics and calculates one minus
the ratio of the difference over the first metrics set.
}
